refactor(rendering): extract EventPresentation into presentation.ts
Phase 2 of the column view. Lifts the view-independent computation out of
renderEvent into a new src/rendering/presentation.ts, so the column view can
derive the identical values instead of recomputing them.

Scope ruling. The plan makes Phase 2 a YAGNI decision point: build a
presentation model only if a named consumer forces it. Two were proposed;
only one is justified.

  EventPresentation - BUILT. Named consumer: the Phase 4 column renderer.
  Spec D1 requires the column view's .event-content to be byte-identical to
  the list's, and that content is driven entirely by EventContentParts, which
  ~120 lines of branching inside renderEvent produce. The only alternatives
  are duplicating those 120 lines per view (guaranteed drift) or sharing the
  builder. This is a real consumer, not speculation.

  DayHeaderModel - NOT BUILT. Phase 1 already solved it without a model:
  renderDateContent takes only primitives and performs its own weekend/today
  colour precedence, leaving renderDateColumn a 4-line composition. Spec D2's
  column header differs from the list header in markup axis only, which is a
  renderer concern. A model here would be abstraction ahead of need. Fold it
  into Phase 4 if that phase ever creates the need.

buildEventPresentation deliberately takes neither the day, the event index,
nor the weather forecasts. Needing none of them is what proves the result is
axis-agnostic; anything positional stays with the container.

Left in render.ts as list-specific or too trivial to drift: dayDate and
isWeekendDay (they feed the date cell, not the event), and the
first/middle/last positional classes (one-line derivations from index).

Two deviations from strict move-discipline, both deliberate:
  - Dropped a dead `tomorrow` local (assigned, mutated, never read). It was
    invisible to tsc because noUnusedLocals is off and to eslint because the
    setDate call counts as a use. Carrying dead code into a new shared module
    would have made it look intentional.
  - Rewrote one comment that was already factually wrong: isEmptyDay adds no
    CSS class, it gates time, countdown and progress.

Verification. The extracted block was diffed against the original and is
byte-identical. The list-view DOM gate passes 86/86 with the snapshot never
regenerated, which is the actual proof of DOM neutrality. Removing the
now-unused FormatUtils and EventUtils imports from render.ts independently
confirms the extraction was complete. All five idiom-guard assertions hold,
including #5 on the ${index === 0} seam, which is the Phase 4 cut point and
was untouched here.

Unlike Phase 1 this moves no markup, so the whitespace trap does not apply.

Copilot-Session: 549dfb35-2cb4-47ba-888c-44d766780ff8
